db: separate RevaultTx helpers in a new bitcointx mod
To keep the schema mod clear and concise
